feat(api): add copilot and issue services for enhanced functionality
