+++ src/qdash/api/lib/sse.py
@@ -1,0 +1,25 @@
+"""Server-Sent Event formatting utilities."""
+
+from __future__ import annotations
+
+import json
+from typing import Any
+
+
+def sse_event(event: str, data: dict[str, Any]) -> str:
+    """Format a Server-Sent Event.
+
+    Parameters
+    ----------
+    event : str
+        The event type name
+    data : dict[str, Any]
+        The event data payload
+
+    Returns
+    -------
+    str
+        Formatted SSE string
+
+    """
+    return f"event: {event}\ndata: {json.dumps(data, ensure_ascii=False)}\n\n"
